Refactor: PostException을 service에서 domain으로 이동
