+++ src/main/java/goorm_runner/backend/mail/presentation/VerificationControllerAdvice.java
@@ -3,7 +3,7 @@ package goorm_runner.backend.mail.presentation;
 import goorm_runner.backend.global.ErrorCode;
 import goorm_runner.backend.global.ErrorResult;
 import goorm_runner.backend.mail.application.exception.MailException;
-import goorm_runner.backend.post.application.post.exception.PostException;
+import goorm_runner.backend.post.domain.exception.PostException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
