+++ src/main/java/goorm_runner/backend/post/application/comment/CommentService.java
@@ -3,10 +3,10 @@ package goorm_runner.backend.post.application.comment;
 import goorm_runner.backend.global.ErrorCode;
 import goorm_runner.backend.post.application.comment.dto.CommentCreateResult;
 import goorm_runner.backend.post.application.comment.dto.CommentUpdateResult;
-import goorm_runner.backend.post.application.post.exception.PostException;
 import goorm_runner.backend.post.domain.CommentRepository;
 import goorm_runner.backend.post.domain.PostRepository;
 import goorm_runner.backend.post.domain.exception.CommentException;
+import goorm_runner.backend.post.domain.exception.PostException;
 import goorm_runner.backend.post.domain.model.Comment;
 import goorm_runner.backend.post.domain.model.Post;
 import lombok.RequiredArgsConstructor;
