+++ src/main/java/goorm_runner/backend/post/application/post/PostReadService.java
@@ -6,8 +6,8 @@ import goorm_runner.backend.member.domain.Member;
 import goorm_runner.backend.member.domain.MemberRepository;
 import goorm_runner.backend.post.application.post.dto.PostReadPageResult;
 import goorm_runner.backend.post.application.post.dto.PostResultOverview;
-import goorm_runner.backend.post.application.post.exception.PostException;
 import goorm_runner.backend.post.domain.PostQueryRepository;
+import goorm_runner.backend.post.domain.exception.PostException;
 import goorm_runner.backend.post.domain.model.Category;
 import goorm_runner.backend.post.domain.model.Post;
 import goorm_runner.backend.postlike.domain.PostLikeRepository;
