+++ src/main/java/goorm_runner/backend/post/application/post/PostService.java
@@ -3,8 +3,8 @@ package goorm_runner.backend.post.application.post;
 import goorm_runner.backend.member.application.exception.MemberException;
 import goorm_runner.backend.member.domain.Member;
 import goorm_runner.backend.member.domain.MemberRepository;
-import goorm_runner.backend.post.application.post.exception.PostException;
 import goorm_runner.backend.post.domain.PostRepository;
+import goorm_runner.backend.post.domain.exception.PostException;
 import goorm_runner.backend.post.domain.model.Category;
 import goorm_runner.backend.post.domain.model.Post;
 import lombok.RequiredArgsConstructor;
