+++ src/main/java/goorm_runner/backend/post/domain/exception/PostException.java
@@ -1,4 +1,4 @@
-package goorm_runner.backend.post.application.post.exception;
+package goorm_runner.backend.post.domain.exception;
 
 import goorm_runner.backend.global.ErrorCode;
 import lombok.Getter;
