+++ src/main/java/goorm_runner/backend/post/presentation/comment/CommentController.java
@@ -9,9 +9,9 @@ import goorm_runner.backend.post.application.comment.dto.CommentCreateResult;
 import goorm_runner.backend.post.application.comment.dto.CommentPageResult;
 import goorm_runner.backend.post.application.comment.dto.CommentReadResult;
 import goorm_runner.backend.post.application.comment.dto.CommentUpdateResult;
-import goorm_runner.backend.post.application.post.exception.PostException;
 import goorm_runner.backend.post.domain.PostQueryRepository;
 import goorm_runner.backend.post.domain.exception.CommentException;
+import goorm_runner.backend.post.domain.exception.PostException;
 import goorm_runner.backend.post.domain.model.Post;
 import goorm_runner.backend.post.presentation.comment.dto.*;
 import lombok.RequiredArgsConstructor;
