+++ src/main/java/goorm_runner/backend/post/presentation/post/PostController.java
@@ -6,8 +6,8 @@ import goorm_runner.backend.member.security.SecurityMember;
 import goorm_runner.backend.post.application.post.PostReadService;
 import goorm_runner.backend.post.application.post.PostService;
 import goorm_runner.backend.post.application.post.dto.PostReadPageResult;
-import goorm_runner.backend.post.application.post.exception.PostException;
 import goorm_runner.backend.post.domain.PostQueryRepository;
+import goorm_runner.backend.post.domain.exception.PostException;
 import goorm_runner.backend.post.domain.model.Category;
 import goorm_runner.backend.post.domain.model.Post;
 import goorm_runner.backend.post.presentation.post.dto.*;
