+++ src/main/java/goorm_runner/backend/post/presentation/post/PostControllerAdvice.java
@@ -2,7 +2,7 @@ package goorm_runner.backend.post.presentation.post;
 
 import goorm_runner.backend.global.ErrorCode;
 import goorm_runner.backend.global.ErrorResult;
-import goorm_runner.backend.post.application.post.exception.PostException;
+import goorm_runner.backend.post.domain.exception.PostException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
