+++ src/test/java/goorm_runner/backend/post/application/post/PostServiceIntegrationTest.java
@@ -4,8 +4,8 @@ import goorm_runner.backend.global.ErrorCode;
 import goorm_runner.backend.member.domain.Member;
 import goorm_runner.backend.member.domain.MemberRepository;
 import goorm_runner.backend.post.application.post.dto.PostReadPageResult;
-import goorm_runner.backend.post.application.post.exception.PostException;
 import goorm_runner.backend.post.domain.PostRepository;
+import goorm_runner.backend.post.domain.exception.PostException;
 import goorm_runner.backend.post.domain.model.Category;
 import goorm_runner.backend.post.domain.model.Post;
 import jakarta.persistence.EntityManager;
